requirements.txt
anthropic>=0.18.0
Pillow>=9.0.0
